requirements.txt
streamlit==1.29.0
python-docx==1.1.0
textstat==0.7.3
PyYAML==6.0.1